refactor: feature 구조 통일
